add the ability to name your booster badge
